Create snapshot tests
