--- src/tests/unitTests/operate.test.js
@@ -1,21 +1,0 @@
-import operate from '../../logic/operate';
-
-describe('operate function operates correctely', () => {
-  test('numbers are summed correctely', () => {
-    const sum = operate(5, 5, '+');
-    expect(sum).toBe('10');
-  });
-  test('numbers are substracted correctely', () => {
-    const substract = operate(5, 5, '-');
-    expect(substract).toBe('0');
-  });
-
-  test('numbers are multiplied correctely', () => {
-    const multiply = operate(5, 5, 'x');
-    expect(multiply).toBe('25');
-  });
-  test('numbers are divided correctely', () => {
-    const divide = operate(5, 5, '÷');
-    expect(divide).toBe('1');
-  });
-});
